requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
requests>=2.25.0
google-generativeai>=0.8.0 